requirements.txt
appdirs==1.4.3
packaging==16.8
PyGithub==1.34
PyJWT==1.5.0
pyparsing==2.2.0
requests==2.13.0
six==1.10.0
slacker==0.9.42
